fix: refactor utils into own module
