+++ .github/workflows/ci.yml
@@ -12,28 +12,28 @@ jobs:
         with:
           # run Cypress via this script
           # and verify that exactly 3 tests are passing
-          command: node . run --passing 3
+          command: node ./bin run --passing 3
 
       - name: Check --min-passing 🧪
         uses: cypress-io/github-action@v2
         with:
           # we have already installed all dependencies above
           install: false
-          command: node . run --min-passing 1
+          command: node ./bin run --min-passing 1
 
       - name: Check other flags are parsed 🧪
         uses: cypress-io/github-action@v2
         with:
           # we have already installed all dependencies above
           install: false
-          command: node . run --passing 3 --spec '**/spec.js'
+          command: node ./bin run --passing 3 --spec '**/spec.js'
 
       - name: Check --failing 🧪
         uses: cypress-io/github-action@v2
         with:
           # we have already installed all dependencies above
           install: false
-          command: node . run --passing 1 --failing 1 --config integrationFolder=cypress/failing
+          command: node ./bin run --passing 1 --failing 1 --config integrationFolder=cypress/failing
 
       - name: --expect example 1 💎
         run: npm run test:expect:1
